move continuityfs -> cmd/continuity/continuityfs
